requirements.txt
altgraph==0.17.3
anyio==3.6.2
APScheduler==3.10.1
boto3==1.26.98
botocore==1.29.98
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==42.0.4
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.109.1
h11==0.14.0
idna==3.7
jmespath==1.0.1
jwt==1.3.1
llvmlite==0.39.1
numba==0.56.4
numpy==1.23.5
pefile==2023.2.7
phonenumbers==8.13.7
pycparser==2.21
pydantic==1.10.7
pyinstaller==5.13.1
pyinstaller-hooks-contrib==2023.2
PyMySQL==1.0.2
pypiwin32==223
pyshorteners==1.0.1
PySide6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pywin32-ctypes==0.2.0
requests==2.31.0
s3transfer==0.6.0
shiboken2==5.15.2.1
six==1.16.0
sniffio==1.3.0
starlette==0.36.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.7
uvicorn==0.21.1
